Modifying code for each activity

+++ orcid-core/src/main/java/org/orcid/core/manager/impl/AffiliationsManagerImpl.java
@@ -16,12 +16,15 @@
  */
 package org.orcid.core.manager.impl;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import javax.annotation.Resource;
 
 import org.orcid.core.adapter.JpaJaxbEducationAdapter;
 import org.orcid.core.adapter.JpaJaxbEmploymentAdapter;
+import org.orcid.core.exception.WrongOwnerException;
 import org.orcid.core.manager.AffiliationsManager;
 import org.orcid.core.manager.NotificationManager;
 import org.orcid.core.manager.OrcidSecurityManager;
@@ -81,7 +84,16 @@ public class AffiliationsManagerImpl implements AffiliationsManager {
     public OrgAffiliationRelationEntity findAffiliationByUserAndId(String userOrcid, Long affiliationId) {
         if (PojoUtil.isEmpty(userOrcid) || affiliationId == null)
             return null;
-        return affiliationsDao.getOrgAffiliationRelation(userOrcid, affiliationId);
+        OrgAffiliationRelationEntity affiliation = affiliationsDao.find(affiliationId);
+        if(affiliation != null) {
+            if(!userOrcid.equals(affiliation.getProfile().getId())) {
+                Map<String, String> params = new HashMap<String, String>();
+                params.put("element", "affiliation");
+                params.put("orcid", userOrcid);
+                throw new WrongOwnerException(params);
+            }
+        }
+        return affiliation;
     }
 
     @Override
@@ -172,6 +184,14 @@ public class AffiliationsManagerImpl implements AffiliationsManager {
     @Override
     public Education updateEducationAffiliation(String orcid, Education education) {
         OrgAffiliationRelationEntity educationEntity = affiliationsDao.getOrgAffiliationRelation(orcid, education.getPutCode());
+        
+        if(educationEntity == null || !orcid.equals(educationEntity.getProfile().getId())) {
+            Map<String, String> params = new HashMap<String, String>();
+            params.put("element", "affiliation");
+            params.put("orcid", orcid);
+            throw new WrongOwnerException(params);
+        }
+                
         Visibility originalVisibility = educationEntity.getVisibility();
         SourceEntity existingSource = educationEntity.getSource();
         orcidSecurityManager.checkSource(existingSource);
@@ -265,6 +285,16 @@ public class AffiliationsManagerImpl implements AffiliationsManager {
     @Override
     public Employment updateEmploymentAffiliation(String orcid, Employment employment) {
         OrgAffiliationRelationEntity employmentEntity = affiliationsDao.getOrgAffiliationRelation(orcid, employment.getPutCode());
+        
+        if(employmentEntity == null || !orcid.equals(employmentEntity.getProfile().getId())) {
+            if(!orcid.equals(employmentEntity.getProfile().getId())) {
+                Map<String, String> params = new HashMap<String, String>();
+                params.put("element", "affiliation");
+                params.put("orcid", orcid);
+                throw new WrongOwnerException(params);
+            }
+        }
+        
         Visibility originalVisibility = employmentEntity.getVisibility();
         SourceEntity existingSource = employmentEntity.getSource();
         orcidSecurityManager.checkSource(existingSource);
@@ -298,9 +328,19 @@ public class AffiliationsManagerImpl implements AffiliationsManager {
     @Override
     public boolean checkSourceAndDelete(String orcid, Long affiliationId) {
         OrgAffiliationRelationEntity affiliationEntity = affiliationsDao.getOrgAffiliationRelation(orcid, affiliationId);
+        
+        if(affiliationEntity == null || !orcid.equals(affiliationEntity.getProfile().getId())) {
+            Map<String, String> params = new HashMap<String, String>();
+            params.put("element", "affiliation");
+            params.put("orcid", orcid);
+            throw new WrongOwnerException(params);
+        }
+                
         orcidSecurityManager.checkSource(affiliationEntity.getSource());
-        notificationManager.sendAmendEmail(orcid, AmendedSection.EMPLOYMENT, createItem(affiliationEntity));
-        return affiliationsDao.removeOrgAffiliationRelation(orcid, affiliationId);
+        boolean result = affiliationsDao.removeOrgAffiliationRelation(orcid, affiliationId);
+        if(result)
+            notificationManager.sendAmendEmail(orcid, AmendedSection.EMPLOYMENT, createItem(affiliationEntity));
+        return result; 
     }
 
     private void setIncomingWorkPrivacy(OrgAffiliationRelationEntity orgAffiliationRelationEntity, ProfileEntity profile) {
